Added automatic mode with PID

+++ src/com/sourcegasm/slothvision/control/Limiter.java
@@ -2,7 +2,7 @@ package com.sourcegasm.slothvision.control;
 
 public class Limiter {
 
-	public static double limitValues(double source, int absLimit) {
+	public static double limitValues(double source, double absLimit) {
 
 		if (source > absLimit)
 			return absLimit;
